first

+++ backend/api/views.py
@@ -2,77 +2,62 @@ from django.http import FileResponse
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
 from recipes.models import Favorite, Recipe, ShoppingCart, Tag, Ingredient
-from rest_framework import filters, status, viewsets, mixins
+from rest_framework import status, viewsets, mixins
 from rest_framework.decorators import action
-from rest_framework.pagination import PageNumberPagination
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 from users.models import FoodgramUser, Subscription
 
 from .serializers import (
-    FavoriteCreteSerializer, IngredientSerializer, RecipeCreateSerializer,
+    FavoriteCreateSerializer, IngredientSerializer, RecipeCreateSerializer,
     RecipeMinifiedSerializer, RecipeSerializer, ShoppingCartCreateSerializer,
     SubscriptionCreateSerializer, SubscriptionSerializer, TagSerializer)
 from .permissions import IsAuthorOrReadOnly
-from .filters import IngredientFilter
+from .filters import IngredientFilter, RecipeFilter
 from .utils import generate_shopping_list_pdf, process_shopping_list
 from api.paginations import CustomPagination
 
 
-class SubscriptionViewSet(viewsets.ViewSet):
+class SubscriptionViewSet(viewsets.GenericViewSet):
     permission_classes = [IsAuthenticated]
     serializer_class = SubscriptionSerializer
+    pagination_class = CustomPagination
+    http_method_names = ('get', 'post', 'delete')
 
-    def get_user(self, pk):
-        return get_object_or_404(FoodgramUser, id=pk)
-
-    def get_serializer(self, *args, **kwargs):
-        return self.serializer_class(*args, **kwargs)
-
-    @action(detail=False, methods=['get'], url_path='subscriptions',
-            url_name='list_subscriptions')
-    def list_subscriptions(self, request):
-        queryset = Subscription.objects.filter(
-            user=request.user).order_by('-pk')
-        paginator = PageNumberPagination()
-        paginated_queryset = paginator.paginate_queryset(queryset, request)
-        serializer = self.get_serializer(paginated_queryset,
-                                         context={'request': request},
-                                         many=True)
+    @action(detail=False, methods=['get'])
+    def subscriptions(self, request):
+        queryset = FoodgramUser.objects.filter(subscribers__user=request.user)
+        paginator = self.pagination_class()
+        page = paginator.paginate_queryset(queryset, request)
+        serializer = SubscriptionSerializer(page, many=True,
+                                            context={'request': request})
         return paginator.get_paginated_response(serializer.data)
 
-    @action(detail=True, methods=['post'], url_path='subscribe',
-            url_name='subscribe')
-    def subscribe(self, request, pk=None):
-        target_user = self.get_user(pk)
+    @action(detail=True, methods=['post', 'delete'])
+    def subscribe(self, request, **kwargs):
+        author = get_object_or_404(FoodgramUser, id=kwargs['pk'])
         serializer = SubscriptionCreateSerializer(
-            data={'author': target_user.id},
+            data={'author': author.id},
             context={'request': request})
         serializer.is_valid(raise_exception=True)
-        subscription = serializer.save(user=request.user)
-        return Response(SubscriptionSerializer(
-            subscription,
-            context={'request': request}).data,
-            status=status.HTTP_201_CREATED)
 
-    @action(detail=True, methods=['delete'], url_path='subscribe',
-            url_name='unsubscribe')
-    def unsubscribe(self, request, pk=None):
-        target_user = self.get_user(pk)
-        serializer = SubscriptionCreateSerializer(
-            data={'author': target_user.id}, context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        subscription = Subscription.objects.get(
-            user=request.user, author=target_user)
-        subscription.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.method == 'POST':
+            subscription = serializer.save(user=request.user)
+            serialized_data = SubscriptionSerializer(
+                author, context={'request': request}).data
+            return Response(serialized_data, status=status.HTTP_201_CREATED)
+
+        if request.method == 'DELETE':
+            subscription = Subscription.objects.get(
+                user=request.user, author=author)
+            subscription.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class TagViewSet(mixins.ListModelMixin,
                  mixins.RetrieveModelMixin,
                  viewsets.GenericViewSet):
-
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
     pagination_class = None
@@ -92,119 +77,56 @@ class RecipeViewSet(ModelViewSet):
     queryset = Recipe.objects.all()
     permission_classes = [IsAuthorOrReadOnly]
     pagination_class = CustomPagination
-    filter_backends = [filters.SearchFilter, filters.OrderingFilter]
+    filter_backends = (DjangoFilterBackend, )
+    filterset_class = RecipeFilter
     http_method_names = ['get', 'post', 'put', 'patch', 'delete']
 
-    def get_queryset(self):
-        queryset = Recipe.objects.select_related('author').prefetch_related(
-            'recipe_ingredients__ingredient', 'tags'
-        )
-
-        author_id = self.request.query_params.get('author', None)
-        if author_id is not None:
-            queryset = queryset.filter(author_id=author_id)
-
-        tags = self.request.query_params.getlist('tags', [])
-        if tags:
-            queryset = queryset.filter(tags__slug__in=tags).distinct()
-
-        is_favorited = self.request.query_params.get('is_favorited', None)
-        if is_favorited is not None:
-            favorited_recipes_ids = self.request.user.favorites_user.all(
-            ).values_list('recipe_id', flat=True)
-            queryset = queryset.filter(id__in=favorited_recipes_ids)\
-                if int(is_favorited)\
-                else queryset.exclude(id__in=favorited_recipes_ids)
-
-        is_in_shopping_cart = self.request.query_params.get(
-            'is_in_shopping_cart', None)
-        if is_in_shopping_cart is not None:
-            shopping_cart_recipe_ids = self.request.user.shopping_cart.all(
-            ).values_list('recipe_id', flat=True)
-            queryset = queryset.filter(id__in=shopping_cart_recipe_ids)\
-                if int(is_in_shopping_cart)\
-                else queryset.exclude(
-                id__in=shopping_cart_recipe_ids)
-
-        return queryset
-
-    @action(detail=True, methods=['post'], url_path='favorite',
-            url_name='add_favorite', permission_classes=[IsAuthenticated])
-    def add_favorite(self, request, pk=None):
-        serializer = FavoriteCreteSerializer(data={'recipe': pk},
-                                             context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        favorite = serializer.save(user=request.user)
-        favorite_data = RecipeMinifiedSerializer(
-            favorite.recipe,
-            context={'request': request}).data
-        return Response(data=favorite_data, status=status.HTTP_201_CREATED)
-
-    @action(detail=True, methods=['delete'], url_path='favorite',
-            url_name='remove_favorite', permission_classes=[IsAuthenticated])
-    def remove_favorite(self, request, pk=None):
-        serializer = FavoriteCreteSerializer(data={'recipe': pk},
-                                             context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        favorite = Favorite.objects.get(
-            user=request.user,
-            recipe=serializer.validated_data['recipe'])
-        favorite.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-    @action(detail=True, methods=["post"], url_path='shopping_cart',
-            url_name='add_to_shopping_cart',
-            permission_classes=[IsAuthenticated])
-    def add_shopping_cart(self, request, pk=None):
-        serializer = ShoppingCartCreateSerializer(data={'recipe': pk},
-                                                  context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        shopping_cart = serializer.save(user=request.user)
-        shopping_cart_data = RecipeMinifiedSerializer(
-            shopping_cart.recipe,
-            context={'request': request}).data
-        return Response(
-            data=shopping_cart_data,
-            status=status.HTTP_201_CREATED)
-
-    @action(detail=True, methods=["delete"], url_path='shopping_cart',
-            url_name='remove_from_shopping_cart',
-            permission_classes=[IsAuthenticated])
-    def remove_shopping_cart(self, request, pk=None):
-        serializer = ShoppingCartCreateSerializer(
-            data={'recipe': pk},
-            context={'request': request})
-        serializer.is_valid(raise_exception=True)
-        favorite = ShoppingCart.objects.get(
-            user=request.user,
-            recipe=serializer.validated_data['recipe'])
-        favorite.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
     def get_serializer_class(self):
-        if self.action == 'create':
-            return RecipeCreateSerializer
-        if self.action == 'update' or self.action == 'partial_update':
+        if self.action in ('create', 'update', 'partial_update'):
             return RecipeCreateSerializer
         return RecipeSerializer
 
-    def perform_create(self, serializer):
-        serializer.save(author=self.request.user)
-        self.request.user.save()
+    @staticmethod
+    def common_action(serializer_class, model, pk, request):
+        serializer = serializer_class(data={'recipe': pk,
+                                            'user': request.user},
+                                      context={'request': request})
+        serializer.is_valid(raise_exception=True)
 
+        if request.method == 'POST':
+            obj = serializer.save(user=request.user)
+            data = RecipeMinifiedSerializer(obj.recipe,
+                                            context={'request': request}).data
+            return Response(data=data, status=status.HTTP_201_CREATED)
 
-class ShoppingCartViewSet(viewsets.ViewSet):
-    permission_classes = [IsAuthenticated]
-    serializer_class = RecipeMinifiedSerializer
-    pagination_class = None
+        if request.method == 'DELETE':
+            obj = model.objects.get(
+                user=request.user,
+                recipe=serializer.validated_data['recipe'])
+            obj.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
-    @action(detail=True, methods=['get'], url_path='shopping_cart',
-            url_name='download_shopping_cart',
+    @action(detail=True, methods=['post', 'delete'], url_path='favorite',
+            url_name='action_with_favorits',
             permission_classes=[IsAuthenticated])
+    def action_with_favorits(self, request, pk=None):
+        return self.common_action(FavoriteCreateSerializer, Favorite,
+                                  pk, request)
+
+    @action(detail=True, methods=['post', 'delete'], url_path='shopping_cart',
+            url_name='shopping_cart_actions',
+            permission_classes=[IsAuthenticated])
+    def shopping_cart_actions(self, request, pk=None):
+        return self.common_action(ShoppingCartCreateSerializer, ShoppingCart,
+                                  pk, request)
+
+    @action(detail=False, methods=['get'], url_path='download_shopping_cart',
+            url_name='download_shopping_cart',
+            permission_classes=[IsAuthenticated],)
     def download_shopping_cart(self, request, file_ext='pdf'):
         user = request.user
         current_user_shopping_list = Recipe.objects.filter(
-            shopping_cart__user=user)
+            shoppingcart__user=user)
         shopping_list_items = process_shopping_list(current_user_shopping_list)
 
         if file_ext == 'pdf':
